Create Programming-Data_Science&ML
Programming for Data Science and Machine Learning, Data science or machine learning
At its core, data science is a field of study that aims to use a scientific approach to extract meaning and insights from data. ... Machine learning, on the other hand, refers to a group of techniques used by data scientists that allow computers to learn from data.